fix: neovim hanging when in command line window
The command line window didn't like that I was using the `:normal`
command for pressing keys. For no apparent reason, this caused it to
hang neovim until ctrl-c was pressed.

The fix was to use `nvim_feedkeys()`, specifically with the `x!` mode.
The documentation doesn't make it clear what this mode does or why it's
required to fix the issue, instead of say the `x` or `n` modes.